+++ flang/lib/Frontend/CMakeLists.txt
@@ -44,6 +44,7 @@ add_flang_library(flangFrontend
 
   LINK_COMPONENTS
   Passes
+  Plugins
   Analysis
   Extensions
   IRPrinter
