+++ llvm/lib/CMakeLists.txt
@@ -44,6 +44,7 @@ add_subdirectory(AsmParser)
 add_subdirectory(LineEditor)
 add_subdirectory(ProfileData)
 add_subdirectory(Passes)
+add_subdirectory(Plugins)
 add_subdirectory(TargetParser)
 add_subdirectory(TextAPI)
 add_subdirectory(Telemetry)
