+++ llvm/lib/Plugins/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_llvm_component_library(LLVMPlugins
+  PassPlugin.cpp
+
+  LINK_COMPONENTS
+  Support
+)
