+++ llvm/unittests/Analysis/PluginInlineOrderAnalysisTest.cpp
@@ -3,7 +3,7 @@
 #include "llvm/Config/config.h"
 #include "llvm/IR/Module.h"
 #include "llvm/Passes/PassBuilder.h"
-#include "llvm/Passes/PassPlugin.h"
+#include "llvm/Plugins/PassPlugin.h"
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/raw_ostream.h"
 #include "llvm/Testing/Support/Error.h"
